fixed throwing retryable exceptions for feign retry

+++ src/main/java/org/gbif/ws/client/ClientErrorDecoder.java
@@ -68,18 +68,18 @@ public class ClientErrorDecoder implements ErrorDecoder {
             : new ValidationException();
       case 429:
         return new RetryableException(
-          response.status(),
-          "Too many requests, please try again later",
-          response.request().httpMethod(),
-          null,
-          response.request());
+            response.status(),
+            "Too many requests, please try again later",
+            response.request().httpMethod(),
+            (Long) null,
+            response.request());
       case 500:
         return new RetryableException(
-          response.status(),
-          "An internal server error occurred, please try again later",
-          response.request().httpMethod(),
-          null,
-          response.request());
+            response.status(),
+            "An internal server error occurred, please try again later",
+            response.request().httpMethod(),
+            (Long) null,
+            response.request());
       case 501:
         return new UnsupportedOperationException(
             message != null ? message : "Method not implement yet");
